Move SinusoidalPositionalEncoding to custom
